Audio fixes in Memory and RF - Cafe
Fixed the fact that the resetting RTPC system made the audio jumpy transitioning from Memory to Knock
Fixed the fact that multiple Journal Scribble SFX would play when multiple journal edits are made simultaneously, by limiting instances of this SFX to 1 in WWise. This bug was encountered at the end of RF - Cafe.